fix: harden renderer/runtime/streaming edge cases from audit
- diff.hpp: write_uint_pos now encodes coords >=10000 correctly (was
  emitting non-digit bytes into the CUP escape); CUP buffer 16->24 so it
  can't overflow.
- cmd.hpp: Cmd::map on Task/IsolatedTask captures mapper+dispatch by
  value, fixing a dangling-reference UAF when a mapped task dispatches
  its callback asynchronously.
- app.hpp: the update() drain loop stops as soon as is_running() flips
  off, so a Cmd::quit() no longer runs the side effects of messages
  queued behind it.
- event_source.hpp: clamp the poll() timeout to [0, INT_MAX] before the
  int narrowing (a >INT_MAX duration could wrap negative = block forever).
- commit.cpp: byte_offset_for_cp caches the achieved cp count, not the
  requested one, so a saturated walk can't poison the resume waypoint.
- canvas.cpp/diff.cpp/serialize.cpp: widen cell-offset multiplies to
  size_t before multiplying (was int*int), matching the file convention.
- edit_tool.hpp: drop the dead `mid.size() - mid.size()` expression.
- test_streaming_cache: tighten kBaselineDrift 3->0; the chunk-drift bug
  is fixed and verified drift-free across an exhaustive chunk-size sweep,
  so the guard now locks it in.

All 38 ctest targets pass.

+++ src/render/canvas.cpp
@@ -553,10 +553,13 @@ void Canvas::fill(Rect region, char32_t ch, uint16_t style_id) {
     // Threshold: 4+ full rows ≈ 320+ cells at width=80.
     const bool full_width = (x0 == 0 && x1 == width_);
     if (full_width && (y1 - y0) >= 4) {
-        simd::streaming_fill(base + y0 * width_, static_cast<std::size_t>((y1 - y0) * width_), packed);
+        simd::streaming_fill(base + static_cast<std::size_t>(y0) * width_,
+                             static_cast<std::size_t>(y1 - y0) * static_cast<std::size_t>(width_),
+                             packed);
     } else {
         for (int y = y0; y < y1; ++y) {
-            std::fill(base + y * width_ + x0, base + y * width_ + x1, packed);
+            std::fill(base + static_cast<std::size_t>(y) * width_ + x0,
+                      base + static_cast<std::size_t>(y) * width_ + x1, packed);
         }
     }
 
@@ -640,7 +643,7 @@ void Canvas::clear_row(int y) noexcept {
 void Canvas::clear_rows(int n) {
     if (n <= 0) return;
     int rows = std::min(n, height_);
-    auto count = static_cast<std::size_t>(rows * width_);
+    auto count = static_cast<std::size_t>(rows) * static_cast<std::size_t>(width_);
     uint64_t blank = default_cell();
     // For small regions (typical inline mode), regular fill keeps L1 warm.
     std::fill(cells_.data(), cells_.data() + count, blank);
